factorize and add front types
