+++ front/types/connectors/admin/cli.ts
@@ -643,6 +643,8 @@ export const ZendeskCommandSchema = t.type({
     t.literal("resync-help-centers"),
     t.literal("resync-brand-metadata"),
     t.literal("sync-ticket"),
+    t.literal("get-retention-period"),
+    t.literal("set-retention-period"),
   ]),
   args: t.type({
     wId: t.union([t.string, t.undefined]),
@@ -653,6 +655,7 @@ export const ZendeskCommandSchema = t.type({
     forceResync: t.union([t.literal("true"), t.undefined]),
     ticketId: t.union([t.number, t.undefined]),
     ticketUrl: t.union([t.string, t.undefined]),
+    retentionPeriodDays: t.union([t.number, t.undefined]),
   }),
 });
 export type ZendeskCommandType = t.TypeOf<typeof ZendeskCommandSchema>;
@@ -688,6 +691,13 @@ export const ZendeskFetchBrandResponseSchema = t.type({
 });
 export type ZendeskFetchBrandResponseType = t.TypeOf<
   typeof ZendeskFetchBrandResponseSchema
+>;
+
+export const ZendeskGetRetentionPeriodResponseSchema = t.type({
+  retentionPeriodDays: t.number,
+});
+export type ZendeskGetRetentionPeriodResponseType = t.TypeOf<
+  typeof ZendeskGetRetentionPeriodResponseSchema
 >;
 /**
  * </Zendesk>
@@ -755,6 +765,7 @@ export const AdminResponseSchema = t.union([
   ZendeskCountTicketsResponseSchema,
   ZendeskFetchBrandResponseSchema,
   ZendeskFetchTicketResponseSchema,
+  ZendeskGetRetentionPeriodResponseSchema,
 ]);
 export type AdminResponseType = t.TypeOf<typeof AdminResponseSchema>;
 /**
